Move to src/antenati package layout with pyproject (PEP 621)
Reshape the project so it is a real installable Python package while
keeping the existing PyInstaller build and the GUI binary name intact.
This is the prerequisite for the PyPI publish step planned for v6.

Layout:
- src/antenati/{cli,downloader,errors,gui,http,iiif}.py     (was the
  flat antenati_*.py files)
- src/antenati/__init__.py       Package metadata (__version__,
                                 __author__, __copyright__, __license__,
                                 __contact__) plus the back-compat
                                 re-exports: AntenatiDownloader,
                                 Downloader, ProgressBar, ThreadError,
                                 DEFAULT_SIZE, DEFAULT_N_THREADS.
- src/antenati/__main__.py       Enables `python -m antenati`.
- antenati_gui.py                Thin PyInstaller launcher at the repo
                                 root (was the GUI itself). Kept here so
                                 the existing `pyinstaller antenati_gui.py`
                                 step and the resulting binary name keep
                                 working byte-for-byte.

pyproject.toml:
- Added [build-system] (hatchling + hatch-vcs) and [project] (PEP 621)
  with the runtime dependencies, classifiers, project URLs and entry
  points (`antenati` CLI, `antenati-gui` GUI).
- __version__ is now dynamic: hatch-vcs derives it from the latest git
  tag at build time; importlib.metadata.version exposes it at runtime.
  A '0.0.0+local' fallback is used when running from a source checkout
  that is not pip-installed.
- src/antenati/_version.py (auto-generated by hatch-vcs) is gitignored.

CI / tooling:
- pythonapp.yml: lint job and live-download job both run
  `pip install [-e] .`; mypy now targets the whole package tree;
  fetch-depth: 0 lets hatch-vcs see the tags. The live-download job
  invokes the installed `antenati` console script.
- pyinstaller.yml: installs the package (`pip install .`) then runs the
  same `pyinstaller antenati_gui.py` invocation. fetch-depth: 0 again.
- .pre-commit-config.yaml: mypy hook now matches src/antenati/ and the
  root launcher.

Tests:
- All imports updated to `from antenati import http as antenati_http`
  etc. The 58-test offline suite passes with `pip install -e .` in the
  venv.
- test_download_run uses `antenati.cli.run_cli` (the function moved out
  of antenati.__init__ since it depends on tqdm).

README:
- Install/run instructions updated to use `pip install .` and the
  installed `antenati`/`python -m antenati` invocations. The CLI flags
  are unchanged so existing -s/-n/-f/-l/-v/--verbose options keep
  working.
