+++ .github/workflows/pyinstaller.yml
@@ -18,6 +18,9 @@ jobs:
     steps:
     - name: Checkout code
       uses: actions/checkout@v4
+      with:
+        # hatch-vcs derives the version from git tags.
+        fetch-depth: 0
 
     - name: Set up Python
       uses: actions/setup-python@v6
@@ -27,7 +30,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r requirements.txt
+        pip install .
         pip install pyinstaller
 
     - name: Build executable with PyInstaller
