+++ .gitignore
@@ -30,6 +30,9 @@ share/python-wheels/
 *.egg
 MANIFEST
 
+# Auto-generated by hatch-vcs at build time
+src/antenati/_version.py
+
 # PyInstaller
 #  Usually these files are written by a python script from a template
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
