+++ README.md
@@ -21,22 +21,25 @@ Copy the link to the first page, and paste it in the Archive URL field of the wi
 ### Requirements
 The software is written in Python 3 and requires Python 3.10 or newer. On Windows the version on the Microsoft Store is fine, on Linux use your distribution package manager.
 
-### Usage
-Open your preferite terminal and change directory to where you've extracted the content of this repo. Then execute the following commands.
+### Install
+From a checkout of this repository:
 
-#### Install the dependencies
-The first time you will have to install the dependencies:
+    pip install .
 
-    pip install -r requirements.txt
+This installs an `antenati` command on your `PATH` (and an `antenati-gui` one for the GUI).
 
-#### Run
-To download the images of a gallery, execute the script passing the URL of a collection you want to download as argument:
+### Run
+To download the images of a gallery, pass the URL of the gallery page:
 
-    python3 antenati.py <URL of the album>
+    antenati <URL of the album>
+
+You can also invoke the package directly without installing the script:
+
+    python3 -m antenati <URL of the album>
 
 The files will be downloaded to a new folder named as *ARCHIVE-PLACE-YEAR-TYPE-ID* of the downloaded archive. For more options, see the help:
 
-    python3 antenati.py -h
+    antenati -h
 
 #### Example:
 In the website, navigate to the archive you want to download. For example, for the people born in Viareggio in 1807 you should find the page:
@@ -45,6 +48,6 @@ In the website, navigate to the archive you want to download. For example, for t
 
 Then, copy the link to the first page, and call the script with that link as argument:
 
-    python3 antenati.py https://antenati.cultura.gov.it/ark:/12657/an_ua19944535/w9DWR8x
+    antenati https://antenati.cultura.gov.it/ark:/12657/an_ua19944535/w9DWR8x
 
 The results will be placed in a folder named *archivio-di-stato-di-lucca-stato-civile-napoleonico-viareggio-1807-nati-19944549*.
