+++ src/antenati/__main__.py
@@ -1,0 +1,5 @@
+"""Allow ``python -m antenati ...`` invocation."""
+
+from antenati.cli import main
+
+main()
