+++ src/antenati/downloader.py
@@ -1,9 +1,9 @@
 """Core download orchestration for the Portale Antenati.
 
 The :class:`Downloader` is the main entry point used by both the CLI
-(:mod:`antenati`) and the GUI (:mod:`antenati_gui`). It composes the
-side-effect free helpers from :mod:`antenati_iiif` with the HTTP session
-built in :mod:`antenati_http`, and runs the per-canvas image downloads in
+(:mod:`antenati.cli`) and the GUI (:mod:`antenati.gui`). It composes the
+side-effect free helpers from :mod:`antenati.iiif` with the HTTP session
+built in :mod:`antenati.http`, and runs the per-canvas image downloads in
 a thread pool.
 
 Keeping this orchestration in its own module makes it possible to embed
@@ -28,9 +28,8 @@ from click import confirm, echo
 from requests import RequestException, Session
 from slugify import slugify
 
-import antenati_http
-import antenati_iiif
-from antenati_errors import AntenatiError, ThreadError
+from antenati import http, iiif
+from antenati.errors import AntenatiError, ThreadError
 
 logger = logging.getLogger(__name__)
 
@@ -64,29 +63,29 @@ class Downloader:
 
     def __init__(self, url: str, first: int, last: int | None):
         self.url = url
-        self.session = antenati_http.build_session()
-        self.archive_id = antenati_iiif.get_archive_id_from_url(url)
+        self.session = http.build_session()
+        self.archive_id = iiif.get_archive_id_from_url(url)
         logger.info('Loading manifest for archive %s', self.archive_id)
         self.manifest = self.__load_manifest()
-        self.canvases = antenati_iiif.slice_canvases(self.manifest, first, last)
+        self.canvases = iiif.slice_canvases(self.manifest, first, last)
         self.dirname = self.__generate_dirname()
         self.gallery_length = len(self.canvases)
         logger.info('Manifest loaded: %d canvases selected', self.gallery_length)
 
     def __load_manifest(self) -> dict[str, Any]:
-        gallery_reply = antenati_http.fetch(self.session, self.url)
-        gallery_charset = antenati_http.get_content_charset(gallery_reply) or 'utf-8'
+        gallery_reply = http.fetch(self.session, self.url)
+        gallery_charset = http.get_content_charset(gallery_reply) or 'utf-8'
         gallery_html = gallery_reply.content.decode(gallery_charset)
-        manifest_url = antenati_iiif.parse_manifest_url_from_html(gallery_html, self.url)
+        manifest_url = iiif.parse_manifest_url_from_html(gallery_html, self.url)
         logger.debug('Manifest URL: %s', manifest_url)
-        manifest_reply = antenati_http.fetch(self.session, manifest_url)
-        manifest_charset = antenati_http.get_content_charset(manifest_reply) or 'utf-8'
+        manifest_reply = http.fetch(self.session, manifest_url)
+        manifest_charset = http.get_content_charset(manifest_reply) or 'utf-8'
         return loads(manifest_reply.content.decode(manifest_charset))
 
     def __generate_dirname(self) -> Path:
-        context = antenati_iiif.get_metadata_value(self.manifest, antenati_iiif.META_CONTEXT)
-        year = antenati_iiif.get_metadata_value(self.manifest, antenati_iiif.META_TITLE)
-        typology = antenati_iiif.get_metadata_value(self.manifest, antenati_iiif.META_TYPOLOGY)
+        context = iiif.get_metadata_value(self.manifest, iiif.META_CONTEXT)
+        year = iiif.get_metadata_value(self.manifest, iiif.META_TITLE)
+        typology = iiif.get_metadata_value(self.manifest, iiif.META_TYPOLOGY)
         return Path(slugify(f'{context}-{year}-{typology}-{self.archive_id}'))
 
     def print_gallery_info(self) -> None:
@@ -115,10 +114,10 @@ class Downloader:
     def __thread_main(self, canvas: dict[str, Any], size: int) -> int:
         label = slugify(canvas['label'])
         try:
-            image_url = antenati_iiif.image_url_for_canvas(canvas)
-            url = antenati_iiif.manipulate_image_url(image_url, size)
-            http_reply = antenati_http.fetch(self.session, url)
-            content_type = antenati_http.get_content_type(http_reply)
+            image_url = iiif.image_url_for_canvas(canvas)
+            url = iiif.manipulate_image_url(image_url, size)
+            http_reply = http.fetch(self.session, url)
+            content_type = http.get_content_type(http_reply)
             extension = guess_extension(content_type)
             if not extension:
                 raise RuntimeError(f'{url}: Unable to guess extension "{content_type}"')
